Add double newline if the message is markup and prefixed.
If the message is prefixed with the sender nick, it will break markup
formatting on the same line. This commit introduces a very rudimentary
markup checker, and if the message is deemed to be markup in those
cases, the space between sender nick and message is replaced by a
double newline.