Fixed typo and added minor clarification.
Fixed typo in the GOPATH link.
While there, I added a sentence of clarification to make it clear that first code block is about installing matterbridge and not related to setting up Go. Having no prior experience with Go, I was a bit confused when I go to the Building section. I believe the new line adds minor but useful clarification.