img: Delete now-unused `slack-setup-*.png` files.
Unused because I've edited the wiki page that previously used these
files [1] so that it uses images hosted on
user-images.githubusercontent.com instead; see discussion in #1490.

Also, this will hopefully make it clearer to people that it's not
necessary to make a PR in order to change the images on a wiki page.
That, in turn, will hopefully make it easier for people to keep the
wiki pages up-to-date.

[1] https://github.com/42wim/matterbridge/wiki/Slack-bot-setup